legacy/dats/requirements.txt
# Core dependencies
celery>=5.3.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0
httpx>=0.25.0
anthropic>=0.40.0
tiktoken>=0.5.0
jinja2>=3.1.0
python-dotenv>=1.0.0
numpy>=1.24.0

# API dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sse-starlette>=1.6.0
python-multipart>=0.0.6
slowapi>=0.1.8
websockets>=12.0

# CLI dependencies
typer[all]>=0.9.0
rich>=13.0.0

# OpenTelemetry dependencies
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-exporter-otlp>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0
opentelemetry-instrumentation-celery>=0.42b0
opentelemetry-instrumentation-httpx>=0.42b0
opentelemetry-instrumentation-logging>=0.42b0

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
black>=23.0.0
ruff>=0.1.0
mypy>=1.0.0